admin-backend/requirements.txt
bokeh==2.0.0
djongo==1.3.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
django-extensions==2.2.8
requests
celery[redis]
mysql-connector-python